refactor(books): move tests to test folder & fix package.json
