+++ test/book/book.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing'
-import { BookService } from '../book.service'
-import { BookDto } from '../dto/book.dto'
-import { PrismaService } from '../../prisma/prisma.service'
+import { BookService } from '../../src/book/book.service'
+import { BookDto } from '../../src/book/dto/book.dto'
+import { PrismaService } from '../../src/prisma/prisma.service'
 
 describe('BookService', () => {
   let service: BookService
